Fix #5487
Don't have image comparison decorator depend on matplotlib.tests, which
may be missing.
