+++ lib/matplotlib/tests/__init__.py
@@ -6,8 +6,7 @@ from matplotlib.externals import six
 import difflib
 import os
 
-from matplotlib import rcParams, rcdefaults, use
-
+from matplotlib.testing import setup
 
 _multiprocess_can_split_ = True
 
@@ -21,39 +20,6 @@ if not os.path.exists(os.path.join(
         'You may need to install matplotlib from source to get the '
         'test data.')
 
-
-def setup():
-    # The baseline images are created in this locale, so we should use
-    # it during all of the tests.
-    import locale
-    import warnings
-    from matplotlib.backends import backend_agg, backend_pdf, backend_svg
-
-    try:
-        locale.setlocale(locale.LC_ALL, str('en_US.UTF-8'))
-    except locale.Error:
-        try:
-            locale.setlocale(locale.LC_ALL, str('English_United States.1252'))
-        except locale.Error:
-            warnings.warn(
-                "Could not set locale to English/United States. "
-                "Some date-related tests may fail")
-
-    use('Agg', warn=False)  # use Agg backend for these tests
-
-    # These settings *must* be hardcoded for running the comparison
-    # tests and are not necessarily the default values as specified in
-    # rcsetup.py
-    rcdefaults()  # Start with all defaults
-    rcParams['font.family'] = 'Bitstream Vera Sans'
-    rcParams['text.hinting'] = False
-    rcParams['text.hinting_factor'] = 8
-
-    # Clear the font caches.  Otherwise, the hinting mode can travel
-    # from one test to another.
-    backend_agg.RendererAgg._fontd.clear()
-    backend_pdf.RendererPdf.truetype_font_cache.clear()
-    backend_svg.RendererSVG.fontd.clear()
 
 
 def assert_str_equal(reference_str, test_str,
